Feat: 상호 상태 확인 인 구현구현
