+++ packages/web/src/apis/updateFriends.ts
@@ -1,0 +1,16 @@
+const BASE_URL = `${process.env.NEXT_PUBLIC_API_URL_V1}/friend`
+
+/** 친구 요청 */
+export async function sendFriendRequest(friendId) {
+  try {
+  } catch {}
+}
+
+/** 친구 요청 취소 */
+export async function cancelFriendRequest(friendId) {}
+
+/** 친구 요청 거절 */
+export async function rejectFriendRequest(friendId) {}
+
+/** 친구 요청 수락 */
+export async function acceptFriendRequest(friendId) {}
